can now add reviews to the database and get all reviews (this page will be for reviews specifically for a single product)

+++ backend/src/models/review.py
@@ -4,7 +4,7 @@ class Review(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     reviewer_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
     product_id = db.Column(db.Integer, db.ForeignKey('product.id'), nullable=False)
-    rating = db.Column(db.Integer, nullable=False)
+    rating = db.Column(db.Float, nullable=False)
     explanation = db.Column(db.String(300))
     review_date = db.Column(db.Date, nullable=False)
